working app

+++ app/src/main/java/com/example/quintin/myfianancer/Objects/Item.java
@@ -19,7 +19,7 @@ public class Item {
         return name;
     }
 
-    public float getPrice() {
+    public int getPrice() {
         return price;
     }
